# services/ekstraksi/openapi.yaml
openapi: 3.1.0
info:
  title: OCR NPWP API
  description: 'OCR service for Indonesian NPWP (tax ID card) documents; ServiceOCR in the pipeline. **Async
    pipeline:** the orchestrator POSTs /v1/ekstraksi/jobs after guardrails passed and gets 202; this service
    runs OCR in the background, stores the result, POSTs a stage callback to the orchestrator, and hands
    the job to the structuring service. **Legacy contract:** generate-request-id -> extract-ocr -> get-ocr-result
    runs the whole chain synchronously and stays until the orchestrator has moved to the async flow. The
    raw OCR step is also exposed as /v1/ekstraksi/extract. All endpoints except /health require an X-API-Key
    header.


    ## Conventions (the same for every nilam-ocr service)


    **Authentication.** Every endpoint except `/health` and `/ready` requires the header `X-API-Key`.

    A missing or wrong key answers `401`. The services are reachable only from inside the cluster.


    **Envelope.** Every JSON response, success or error, has the same shape:

    `{status_code, status_desc, message, data, errors, request_id}`. `status_code` always equals the HTTP

    status. On success `errors` is null; on error `data` is null.


    **Errors.** `errors` is machine-readable: `VALIDATION_ERROR` for `422`, otherwise the same text as

    `message`. `400` = the request or the document is unusable (do not retry unchanged). `401` = API key.

...

...

...


...
    no

...

    without it the service generates one.


...

...

...

...

...

...

...
servers:
...
paths:
...
  /v1/ekstraksi/jobs:
    post:
      tags:
      - Pipeline
      summary: Start the pipeline for a document (OCR stage)
      description: '**Step 2 of the pipeline, asynchronous: the entry point of the OCR -> structuring
        -> scoring chain.** Call it after `POST /v1/guardrails/check` answered `passed: true`.


        Records the job (`ocr.jobs`, idempotent per request_id), answers **202 immediately**, then in
        the background: reads the document (`file`, or downloads `file_url`), runs OCR, stores the result
        (`ocr.results`), POSTs the `OCR` callback, and hands the job to the structuring service, which
        hands it to scoring. The caller does nothing more: it receives one callback per stage, and the
        final result in the `SCORING` callback (see *Webhooks*).


        **Document.** Send `file` (multipart) or `file_url`, exactly one. JPEG, PNG or PDF, at most 5
        MB. `file_url` is downloaded in the background, so make a presigned URL live longer than the worst
        queueing time; an expired or unreachable URL becomes a `FAILED` job, not a `4xx`.


        **Idempotency.** The same request_id again answers `202` with `duplicate: true` and does not run
        OCR twice, unless the earlier attempt `FAILED`, in which case it is run again.'
      operationId: submitOcrJob
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_submitOcrJob'
        required: true
      responses:
        '202':
          description: The job was accepted (or already existed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobAcceptedResponse'
              examples:
                accepted:
                  summary: New job
                  value:
                    status_code: 202
                    status_desc: Accepted
                    message: Accepted
                    data:
                      request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
                      stage: OCR
                      status: PROCESSING
                      duplicate: false
                    request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
                duplicate:
                  summary: 'Same request_id sent again: nothing is re-run, `status` is the existing job''s
                    status'
                  value:
                    status_code: 202
                    status_desc: Accepted
                    message: Accepted
                    data:
                      request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
                      stage: OCR
                      status: DONE
                      duplicate: true
                    request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
        '400':
          description: Neither or both of file / file_url, or `guardrails` is not a JSON object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status_code: 400
                status_desc: Bad Request
                message: Send exactly one of file or file_url
                errors: Send exactly one of file or file_url
        '401':
          description: Missing or invalid X-API-Key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status_code: 401
                status_desc: Unauthorized
                message: Invalid or missing API key
                errors: Invalid or missing API key
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status_code: 422
                status_desc: Unprocessable Entity
                message: 'body.request_id: Field required'
                errors: VALIDATION_ERROR
      security:
      - APIKeyHeader: []
...
  /v1/generate-request-id:
    post:
      tags:
      - NPWP OCR
      summary: 'Legacy: mint a request_id'
      description: '**Legacy synchronous contract**, kept until the orchestrator has moved to the asynchronous
        pipeline (`POST /v1/ekstraksi/jobs`). Mints a request_id that must be used once in POST /v1/extract-ocr.
        In the asynchronous pipeline the orchestrator mints the request_id itself.'
      operationId: generateRequestId
      responses:
        '200':
          description: A new request_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateRequestIdResponse'
              examples:
                minted:
                  summary: Use it once in POST /v1/extract-ocr
                  value:
                    status_code: 200
                    status_desc: OK
                    message: Success
                    data:
                      request_id: OCR_9cb01af2-493d-446d-b191-af120333f6d0
                    request_id: OCR_9cb01af2-493d-446d-b191-af120333f6d0
        '401':
          description: Missing or invalid X-API-Key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status_code: 401
                status_desc: Unauthorized
                message: Invalid or missing API key
                errors: Invalid or missing API key
      deprecated: true
      security:
      - APIKeyHeader: []
  /v1/extract-ocr:
    post:
      tags:
      - NPWP OCR
      summary: 'Legacy: run the whole chain synchronously'
      description: '**Legacy synchronous contract**, kept until the orchestrator has moved to the asynchronous
        pipeline (`POST /v1/ekstraksi/jobs`). Runs the full pipeline (guardrails service -> this service''s
        OCR -> structuring service -> scoring service) on the uploaded NPWP image for the given request_id
        (from /v1/generate-request-id). A request_id can only be submitted once.


        Send the image as `file`, or send `file_url` and this service fetches it itself; exactly one of
        the two.'
      operationId: extractOcr
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_extractOcr'
        required: true
      responses:
        '200':
          description: The document was read
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtractOcrResponse'
              examples:
                person:
                  summary: A person's card
                  value:
                    status_code: 200
                    status_desc: OK
                    message: Success
                    data:
                      nomor_npwp:
                        value: 12.345.678.9-012.345
                        confidence: 0.9992
                      nama:
                        value: BUDI SANTOSO
                        confidence: 0.9773
                      nama_badan:
                        confidence: 0.0
                    request_id: OCR_9cb01af2-493d-446d-b191-af120333f6d0
                    guardrails: 0.9904
        '400':
          description: Unknown request_id, bad file, or rejected by guardrails
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status_code: 400
                status_desc: Bad Request
                message: 'Document rejected by guardrails: 1/1 page(s) rejected (confidence 0.88)'
                errors: 'Document rejected by guardrails: 1/1 page(s) rejected (confidence 0.88)'
        '401':
          description: Missing or invalid X-API-Key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status_code: 401
                status_desc: Unauthorized
                message: Invalid or missing API key
                errors: Invalid or missing API key
        '409':
          description: request_id was already processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status_code: 409
                status_desc: Conflict
                message: request_id OCR_9cb01af2-493d-446d-b191-af120333f6d0 has already been processed
                errors: request_id OCR_9cb01af2-493d-446d-b191-af120333f6d0 has already been processed
        '500':
          description: OCR engine or a downstream stage failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status_code: 500
                status_desc: Internal Server Error
                message: Internal server error while processing OCR
                errors: Internal server error while processing OCR
        '503':
          description: OCR model or a downstream stage unreachable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status_code: 503
                status_desc: Service Unavailable
                message: ekstraksi OCR model is unavailable
                errors: ekstraksi OCR model is unavailable
        '504':
          description: OCR model or a downstream stage did not answer in time
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status_code: 504
                status_desc: Gateway Timeout
                message: ekstraksi OCR model timed out after 30.0s
                errors: ekstraksi OCR model timed out after 30.0s
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status_code: 422
                status_desc: Unprocessable Entity
                message: 'body.request_id: Field required'
                errors: VALIDATION_ERROR
      deprecated: true
      security:
      - APIKeyHeader: []
  /v1/get-ocr-result/{request_id}:
    get:
      tags:
      - NPWP OCR
      summary: 'Legacy: status and result of an extract-ocr request'
      description: '**Legacy synchronous contract**, kept until the orchestrator has moved to the asynchronous
        pipeline (`POST /v1/ekstraksi/jobs`). Current status and, once completed, the result of a request_id
        minted by this service.'
      operationId: getOcrResult
      deprecated: true
      security:
      - APIKeyHeader: []
      parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: string
          title: Request Id
      responses:
        '200':
          description: The request_id exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetOcrResultResponse'
              examples:
                completed:
                  summary: Finished
                  value:
                    status_code: 200
                    status_desc: OK
                    message: Success
                    data:
                      request_id: OCR_9cb01af2-493d-446d-b191-af120333f6d0
                      status: completed
                      result:
                        nomor_npwp:
                          value: 12.345.678.9-012.345
                          confidence: 0.9992
                        nama:
                          value: BUDI SANTOSO
                          confidence: 0.9773
                        nama_badan:
                          confidence: 0.0
                      created_at: '2026-03-06T07:01:59.976912+00:00'
                      updated_at: '2026-03-06T07:07:04.809697+00:00'
                    request_id: OCR_9cb01af2-493d-446d-b191-af120333f6d0
                    guardrails: 0.9904
                pending:
                  summary: Minted, but extract-ocr has not been called (or is still running)
                  value:
                    status_code: 200
                    status_desc: OK
                    message: Success
                    data:
                      request_id: OCR_9cb01af2-493d-446d-b191-af120333f6d0
                      status: pending
                      created_at: '2026-03-06T07:01:59.976912+00:00'
                      updated_at: '2026-03-06T07:01:59.976912+00:00'
                    request_id: OCR_9cb01af2-493d-446d-b191-af120333f6d0
        '401':
          description: Missing or invalid X-API-Key
          content:
            application/json:
              example:
                status_code: 401
                status_desc: Unauthorized
                message: Invalid or missing API key
                errors: Invalid or missing API key
                request_id: OCR_9cb01af2-493d-446d-b191-af120333f6d0
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: request_id was never generated by this service
          content:
            application/json:
              example:
                status_code: 404
                status_desc: Not Found
                message: 'No data found for request_id: OCR_9cb01af2-493d-446d-b191-af120333f6d0'
                errors: 'No data found for request_id: OCR_9cb01af2-493d-446d-b191-af120333f6d0'
                request_id: OCR_9cb01af2-493d-446d-b191-af120333f6d0
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              example:
                status_code: 422
                status_desc: Unprocessable Entity
                message: 'path.request_id: Field required'
                errors: VALIDATION_ERROR
                request_id: OCR_9cb01af2-493d-446d-b191-af120333f6d0
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    Body_extractOcr:
      properties:
        request_id:
          type: string
          title: Request Id
          description: request_id from /v1/generate-request-id
        file:
          anyOf:
          - type: string
            format: binary
          - type: string
          - type: 'null'
          title: File
          description: Document image (JPEG/PNG/PDF). Omit when sending file_url.
        file_url:
          anyOf:
          - type: string
          - type: 'null'
          title: File Url
          description: URL this service fetches the document image from (e.g. a presigned MinIO GET).
            Omit when uploading file.
      type: object
      required:
      - request_id
      title: Body_extractOcr
...
    Body_submitOcrJob:
      properties:
        request_id:
          type: string
          title: Request Id
          description: request_id minted by the orchestrator
          examples:
          - REQ_9cb01af2-493d-446d-b191-af120333f6d0
        document_type:
          type: string
          title: Document Type
          description: Document type chosen by the client. Only `npwp` is supported
          default: npwp
          examples:
          - npwp
        guardrails:
          anyOf:
          - type: string
          - type: 'null'
          title: Guardrails
          description: 'The `data` object of `POST /v1/guardrails/check`, serialised as a JSON string.
            Forwarded down the chain: scoring uses `document.confidence`, and the final result returns
            it unchanged. Optional, but without it the scoring model works with one input missing'
          examples:
          - '{"passed": true, "reason": null, "document": {"verdict": "accepted", "confidence": 0.9821,
            "n_pages": 1, "n_approve": 1, "n_reject": 0}, "pages": [{"page_index": 0, "proba_approve":
            0.9821, "proba_reject": 0.0179, "verdict": "accepted"}]}'
        file:
          anyOf:
          - type: string
            format: binary
          - type: string
          - type: 'null'
          title: File
          description: Document image (JPEG/PNG/PDF). Omit when sending file_url.
        file_url:
          anyOf:
          - type: string
          - type: 'null'
          title: File Url
          description: URL this service fetches the document image from (e.g. a presigned MinIO GET).
            Omit when uploading file.
      type: object
      required:
      - request_id
      title: Body_submitOcrJob
...
    ErrorResponse:
      properties:
        status_code:
          type: integer
          title: Status Code
          description: Same as the HTTP status code
          examples:
          - 400
        status_desc:
          type: string
          title: Status Desc
          description: Reason phrase of `status_code`
          examples:
          - Bad Request
        message:
          type: string
          title: Message
          description: Human-readable cause; safe to log
          examples:
          - Uploaded file is empty
        data:
          type: 'null'
          title: Data
          description: Always null on error
        errors:
          type: string
          title: Errors
          description: Machine-readable cause. `VALIDATION_ERROR` for 422 (branch on this, the message
            names a different field every time); for every other status it repeats `message`
          examples:
          - Uploaded file is empty
        request_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Request Id
          description: Present when the error is tied to a known request_id
          examples:
          - REQ_9cb01af2-493d-446d-b191-af120333f6d0
      type: object
      required:
      - status_code
      - status_desc
      - message
      - errors
      title: ErrorResponse
    ExtractOcrResponse:
      properties:
        status_code:
          type: integer
          title: Status Code
          description: Same as the HTTP status code
          default: 200
          examples:
          - 200
        status_desc:
          type: string
          title: Status Desc
          description: Reason phrase of `status_code`
          default: OK
          examples:
          - OK
        message:
          type: string
          title: Message
          description: Human-readable outcome
          default: Success
          examples:
          - Success
        errors:
          type: 'null'
          title: Errors
          description: Always null on success
        request_id:
          type: string
          title: Request Id
          description: The request_id this response belongs to
          examples:
          - OCR_9cb01af2-493d-446d-b191-af120333f6d0
        data:
          $ref: '#/components/schemas/NpwpFields'
        guardrails:
          anyOf:
          - type: number
            maximum: 1.0
            minimum: 0.0
          - type: 'null'
          title: Guardrails
          description: 'Overall document score (0..1) from the legacy heuristic: OCR confidence per field
            combined with format checks. NOT produced by the ML team''s trust model; the async pipeline
            reports per-field confidences instead.'
          examples:
          - 0.98
      type: object
      required:
      - request_id
      - data
      title: ExtractOcrResponse
...
    GenerateRequestIdData:
      properties:
        request_id:
          type: string
          title: Request Id
          description: Use it once in POST /v1/extract-ocr
          examples:
          - OCR_9cb01af2-493d-446d-b191-af120333f6d0
      type: object
      required:
      - request_id
      title: GenerateRequestIdData
    GenerateRequestIdResponse:
      properties:
        status_code:
          type: integer
          title: Status Code
          description: Same as the HTTP status code
          default: 200
          examples:
          - 200
        status_desc:
          type: string
          title: Status Desc
          description: Reason phrase of `status_code`
          default: OK
          examples:
          - OK
        message:
          type: string
          title: Message
          description: Human-readable outcome
          default: Success
          examples:
          - Success
        errors:
          type: 'null'
          title: Errors
          description: Always null on success
        request_id:
          type: string
          title: Request Id
          description: The request_id this response belongs to
          examples:
          - OCR_9cb01af2-493d-446d-b191-af120333f6d0
        data:
          $ref: '#/components/schemas/GenerateRequestIdData'
      type: object
      required:
      - request_id
      - data
      title: GenerateRequestIdResponse
...
    GetOcrResultResponse:
      properties:
        status_code:
          type: integer
          title: Status Code
          description: Same as the HTTP status code
          default: 200
          examples:
          - 200
        status_desc:
          type: string
          title: Status Desc
          description: Reason phrase of `status_code`
          default: OK
          examples:
          - OK
        message:
          type: string
          title: Message
          description: Human-readable outcome
          default: Success
          examples:
          - Success
        errors:
          type: 'null'
          title: Errors
          description: Always null on success
        request_id:
          type: string
          title: Request Id
          description: The request_id this response belongs to
          examples:
          - OCR_9cb01af2-493d-446d-b191-af120333f6d0
        data:
          $ref: '#/components/schemas/GetOcrResultData'
        guardrails:
          anyOf:
          - type: number
            maximum: 1.0
            minimum: 0.0
          - type: 'null'
          title: Guardrails
          description: 'Overall document score (0..1) from the legacy heuristic: OCR confidence per field
            combined with format checks. NOT produced by the ML team''s trust model; the async pipeline
            reports per-field confidences instead. Null until `status` is `completed`.'
          examples:
          - 0.98
      type: object
      required:
      - request_id
      - data
      title: GetOcrResultResponse
...
    JobAccepted:
      properties:
        request_id:
          type: string
          title: Request Id
          description: Echo of the submitted request_id
          examples:
          - REQ_9cb01af2-493d-446d-b191-af120333f6d0
        stage:
          type: string
          enum:
          - OCR
          - STRUCTURING
          - SCORING
          title: Stage
          description: Pipeline stage that accepted the job
          examples:
          - OCR
        status:
          type: string
          enum:
          - PROCESSING
          - DONE
          - FAILED
          title: Status
          description: '`PROCESSING` for a newly accepted job. When `duplicate` is true: the current status
            of the job that already exists for this request_id'
          examples:
          - PROCESSING
        duplicate:
          type: boolean
          title: Duplicate
          description: 'true: this request_id was already accepted before and the work is NOT run again
            (idempotent). Exception: a job whose status is `FAILED` is re-claimed and run again, and answers
            `duplicate: false`'
          default: false
          examples:
          - false
      type: object
      required:
      - request_id
      - stage
      - status
      title: JobAccepted
    JobAcceptedResponse:
      properties:
        status_code:
          type: integer
          title: Status Code
          description: Same as the HTTP status code
          default: 202
          examples:
          - 202
        status_desc:
          type: string
          title: Status Desc
          description: Reason phrase of `status_code`
          default: Accepted
          examples:
          - Accepted
        message:
          type: string
          title: Message
          description: Human-readable outcome
          default: Accepted
          examples:
          - Accepted
        errors:
          type: 'null'
          title: Errors
          description: Always null on success
        request_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Request Id
          description: 'The request_id this response belongs to: the one sent in the form / body / path,
            otherwise the `X-Request-ID` header, otherwise one generated by the service'
          examples:
          - REQ_9cb01af2-493d-446d-b191-af120333f6d0
        data:
          $ref: '#/components/schemas/JobAccepted'
      type: object
      required:
      - data
      title: JobAcceptedResponse
    NpwpFields:
      properties:
        nomor_npwp:
          $ref: '#/components/schemas/Confidence'
          description: 15-digit number as `XX.XXX.XXX.X-XXX.XXX`, or a 16-digit NIK-based number as 16
            plain digits
          examples:
          - confidence: 0.99
            value: 12.345.678.9-012.345
        nama:
          $ref: '#/components/schemas/Confidence'
          description: Taxpayer name on a person's card
          examples:
          - confidence: 0.97
            value: BUDI SANTOSO
        nama_badan:
          $ref: '#/components/schemas/Confidence'
          description: Registered name on a company's card (PT, CV, ...); null on a person's card
          examples:
          - confidence: 0.0
      type: object
      required:
      - nomor_npwp
      - nama
      - nama_badan
      title: NpwpFields
...
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      description: API key for this service
      in: header
      name: X-API-Key
...